Revert "perf(ad/smartbanner): move component to function and use dynamic import to load only if needed the svg"
