+++ components/ad/smartbanner/src/index.js
@@ -1,71 +1,89 @@
 import PropTypes from 'prop-types'
-import React from 'react'
+import React, {Component} from 'react'
 import IconCloseDefault from '@schibstedspain/sui-svgiconset/lib/X'
+import RatingStar from './rating-star'
 import cx from 'classnames'
 
-import RatingStar from './RatingStar'
+class AdSmartbanner extends Component {
+  _handleClick = event => {
+    const {onClick} = this.props
+    event.preventDefault()
+    onClick()
+  }
 
-function AdSmartbanner({
-  buttonText,
-  customRatingIcons,
-  icon: IconClose,
-  imageUrl,
-  onClick,
-  onClose,
-  ratingMax,
-  ratingValue,
-  staticPosition,
-  text,
-  title
-}) {
-  const className = cx('sui-AdSmartbanner', {
-    'is-static': staticPosition
-  })
+  _handleClose = event => {
+    const {onClose} = this.props
+    event.preventDefault()
+    onClose()
+  }
 
-  return (
-    <div className={className}>
-      <button className="sui-AdSmartbanner-buttonClose" onClick={onClose}>
-        <IconClose svgClass="sui-AdSmartbanner-buttonCloseIcon" />
-      </button>
-      <div className="sui-AdSmartbanner-primary">
-        <img alt="logo" src={imageUrl} className="sui-AdSmartbanner-logo" />
+  render() {
+    const {
+      imageUrl,
+      title,
+      text,
+      buttonText,
+      staticPosition,
+      ratingValue,
+      ratingMax,
+      customRatingIcons,
+      icon: IconClose
+    } = this.props
+    const className = cx('sui-AdSmartbanner', {
+      'is-static': staticPosition
+    })
+
+    return (
+      <div className={className}>
+        <button
+          className="sui-AdSmartbanner-buttonClose"
+          onClick={this._handleClose}
+        >
+          <IconClose svgClass="sui-AdSmartbanner-buttonCloseIcon" />
+        </button>
+        <div className="sui-AdSmartbanner-primary">
+          <img alt="logo" src={imageUrl} className="sui-AdSmartbanner-logo" />
+        </div>
+        <div className="sui-AdSmartbanner-secondary">
+          <h3 className="sui-AdSmartbanner-title">{title}</h3>
+          <p className="sui-AdSmartbanner-text">{text}</p>
+          {ratingValue !== null && (
+            <RatingStar
+              ratingValue={ratingValue}
+              ratingMax={ratingMax}
+              icons={customRatingIcons}
+            />
+          )}
+        </div>
+        <button
+          className="sui-AdSmartbanner-buttonInstall"
+          onClick={this._handleClick}
+        >
+          {buttonText}
+        </button>
       </div>
-      <div className="sui-AdSmartbanner-secondary">
-        <h3 className="sui-AdSmartbanner-title">{title}</h3>
-        <p className="sui-AdSmartbanner-text">{text}</p>
-        {ratingValue !== null && (
-          <RatingStar
-            ratingValue={ratingValue}
-            ratingMax={ratingMax}
-            icons={customRatingIcons}
-          />
-        )}
-      </div>
-      <button className="sui-AdSmartbanner-buttonInstall" onClick={onClick}>
-        {buttonText}
-      </button>
-    </div>
-  )
+    )
+  }
 }
 
 AdSmartbanner.propTypes = {
-  buttonText: PropTypes.string.isRequired,
-  customRatingIcons: PropTypes.object,
-  icon: PropTypes.func,
-  imageUrl: PropTypes.string.isRequired,
   onClick: PropTypes.func.isRequired,
   onClose: PropTypes.func.isRequired,
-  ratingMax: PropTypes.number,
-  ratingValue: PropTypes.number,
-  staticPosition: PropTypes.bool,
+  icon: PropTypes.func,
+  imageUrl: PropTypes.string.isRequired,
+  title: PropTypes.string.isRequired,
   text: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired
+  buttonText: PropTypes.string.isRequired,
+  staticPosition: PropTypes.bool,
+  customRatingIcons: PropTypes.object,
+  ratingValue: PropTypes.number,
+  ratingMax: PropTypes.number
 }
 
 AdSmartbanner.defaultProps = {
-  icon: IconCloseDefault,
+  staticPosition: false,
   ratingValue: null,
-  staticPosition: false
+  icon: IconCloseDefault
 }
 
 AdSmartbanner.displayName = 'AdSmartbanner'
