+++ components/ad/smartbanner/src/rating-star.js
